Clean up existing folders
