--- pkg/data/file_manager_test.go
@@ -1,1 +1,0 @@
-package pkg
